+++ src/ts/types/beatmap/mod.ts
@@ -2,3 +2,4 @@ export * from './shared/mod';
 export * as external from './external/mod';
 export * as v2 from './v2/mod';
 export * as v3 from './v3/mod';
+export * as wrapper from './wrapper/mod';